moving select choice to input
